fix: moved data for PhotoPage to /data and removed the redundant /events folder in /data
