+++ gatsby-node.js
@@ -8,8 +8,8 @@ const path = require('path');
 const { createFilePath } = require('gatsby-source-filesystem');
 const dayjs = require('dayjs');
 
-const events = require('./src/data/events/events');
-const highlightedEvents = require('./src/data/events/highlights');
+const events = require('./src/data/events');
+const highlightedEvents = require('./src/data/highlights');
 
 // convert windows to linux path
 const postDirectory = path.join(__dirname, 'posts').replace(/\\/g, '/');
